Move authentication and permission classes to core
