+++ core/authentication.py
@@ -1,4 +1,9 @@
 from rest_framework import authentication, exceptions
+from rest_framework.authentication import BasicAuthentication
+
+from django.conf import settings
+from django.contrib.auth.models import User
+from django.utils.translation import ugettext_lazy as _
 
 from supplier import helpers
 
@@ -36,6 +41,13 @@ class SessionAuthenticationSSO(authentication.BaseAuthentication):
 
 
 class Oauth2AuthenticationSSO(authentication.BaseAuthentication):
+    """
+    Clients should authenticate by passing an oauth2 bearer token in the
+    "Authorization" HTTP header, prepended with the string "Bearer ".
+    For example:
+        Authorization: Bearer th38ear3rt0k3n
+    """
+
     message_invalid_session = 'Invalid bearer token'
     message_bad_format = 'Invalid bearer header.'
     keyword = 'Bearer'
@@ -61,3 +73,25 @@ class Oauth2AuthenticationSSO(authentication.BaseAuthentication):
 
     def authenticate_header(self, request):
         return self.keyword
+
+
+class GeckoBasicAuthentication(BasicAuthentication):
+    """Authentication class that uses a username and password setting
+    instead of a django user
+
+    DRF's BasicAuthentication class uses the auth.User model for
+    authentication. Credentials created for gecko could therefore
+    be used to access other parts of the site. The purpose of this
+    class is to give gecko access to just views that use this
+    class and absolutely nothing else."""
+
+    def authenticate_credentials(self, userid, password):
+        username_invalid = (userid != settings.GECKO_API_KEY)
+        password_invalid = (password != settings.GECKO_API_PASS)
+        if username_invalid or password_invalid:
+            raise exceptions.AuthenticationFailed(
+                _('Invalid username/password.'))
+        else:
+            user = User(username=userid)
+
+        return (user, None)
